requirements.txt
# Core Machine Learning and Deep Learning
tensorflow
keras
numpy
pandas

# Image Processing
Pillow

# Visualization and Analysis
matplotlib
seaborn
scikit-learn

# Deployment and Data Handling
streamlit
split-folders